Keybind Settings
Created buttons in the ControlsPanel that let player set their own keybinds using the new Unity Input System with the "Input System_Actions" asset. Had to modify PlayerMovement and PlayerVisuals scripts to replace old input system with the new customizable one.

As we add more controls to the game, they can be added as buttons in the ControlsPanel, but someone might have to make a scrollbar because there isn't much room.

In the future, we also need to save settings as PlayerPrefs.